requirements.txt
# RAG Research Assistant — production stack
# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# LangChain orchestration
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector store
chromadb>=0.5.0

# PDF
pypdf>=5.0.0

# Config
python-dotenv>=1.0.0

# Frontend
streamlit>=1.40.0
httpx>=0.27.0

# Typing / quality
pydantic>=2.0.0
